move the exception to the root path
